Object oriented approach
