+++ src/structures/board.rs
@@ -1,0 +1,37 @@
+use crate::battlesnake::Battlesnake;
+use crate::coordinate::Coordinate;
+
+#[derive(Clone, Debug)]
+pub struct Board {
+    height: i32,
+    width: i32,
+    food: Vec<Coordinate>,
+    hazards: Vec<Coordinate>,
+    snakes: Vec<Battlesnake>
+}
+
+impl Board {
+    pub fn new(height: i32, width: i32, food: Vec<Coordinate>, hazards: Vec<Coordinate>, snakes: Vec<Battlesnake>) -> Board {
+        Board {height, width, food, hazards, snakes}
+    }
+
+    pub fn get_height(&self) -> i32 {
+        self.height
+    }
+
+    pub fn get_width(&self) -> i32 {
+        self.width
+    }
+
+    pub fn get_food(&mut self) -> &mut Vec<Coordinate> {
+        &mut self.food
+    }
+
+    pub fn _get_hazards(&self) -> &Vec<Coordinate> {
+        &self.hazards
+    }
+
+    pub fn get_snakes(&mut self) -> &mut Vec<Battlesnake> {
+        &mut self.snakes
+    }
+}
