+++ src/structures/game.rs
@@ -1,0 +1,28 @@
+use serde::{Deserialize, Serialize};
+
+use crate::ruleset::Ruleset;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Game {
+    id: String,
+    ruleset: Ruleset,
+    timeout: i32
+}
+
+impl Game {
+    pub fn _new(id: String, ruleset: Ruleset, timeout: i32) -> Game {
+        Game {id, ruleset, timeout}
+    }
+
+    pub fn get_id(&self) -> &String {
+        &self.id
+    }
+
+    pub fn _get_ruleset(&self) -> &Ruleset {
+        &self.ruleset
+    }
+
+    pub fn _get_timeout(&self) -> i32 {
+        self.timeout
+    }
+}
