+++ src/structures/mod.rs
@@ -1,0 +1,5 @@
+pub mod battlesnake;
+pub mod board;
+pub mod coordinate;
+pub mod game;
+pub mod ruleset;
